cmd files changed based on new repositories

+++ cmd/Category.go
@@ -5,26 +5,39 @@ import (
 	"math/rand/v2"
 	"todo/model"
 	"todo/repository"
+	"todo/utilities"
 )
 
 type CategoryCommand struct {
 	repo repository.CategoryRepository
+	ioUtil utilities.IO
 }
+
 
 
 func NewCategoryCommand(repo repository.CategoryRepository) CategoryCommand {
-	return CategoryCommand{repo}
+	return CategoryCommand{repo: repo, ioUtil: utilities.NewIO()}
 }
 
-func (cmd CategoryCommand) CreateCategory() {
+func (cmd *CategoryCommand) CreateCategory() {
 	var category model.Category = model.Category{}
 	fmt.Println("Create Category Command ...")
 	
 	fmt.Println("Please enter the title of the category : ")
-	fmt.Scanln(&category.Title)
+	categoryTitle, readCatTitleErr := cmd.ioUtil.ReadStr()
+	if(readCatTitleErr != nil) {
+		fmt.Println(readCatTitleErr)
+		return
+	}
+	category.Title = categoryTitle
 
 	fmt.Println("Please enter the color of the category :")
-	fmt.Scanln(&category.Color)
+	categoryColor, readCatColorErr := cmd.ioUtil.ReadStr()
+	if(readCatColorErr != nil) {
+		fmt.Println(readCatColorErr)
+		return
+	}
+	category.Color = categoryColor
 	category.Id = rand.IntN(100)
 
 	err := cmd.repo.CreateCategory(category)
@@ -33,7 +46,7 @@ func (cmd CategoryCommand) CreateCategory() {
 	} 
 	fmt.Println("Category created successfully")
 }
-func (cmd CategoryCommand) ListAllCategories() {
+func (cmd *CategoryCommand) ListAllCategories() {
 	fmt.Println("List All Categories Command ...")
 	categories, err := cmd.repo.ListAllCategories()
 	if err != nil {
@@ -42,30 +55,46 @@ func (cmd CategoryCommand) ListAllCategories() {
 	fmt.Println("List of all categories : ")
 	fmt.Println(categories)
 }
-func (cmd CategoryCommand) GetCategory() {
+func (cmd *CategoryCommand) GetCategory() {
 	fmt.Println("Get Category Command ...")
-	var categoryId int
 	fmt.Println("Please enter the id of the category : ")
-	fmt.Scanf("%d", &categoryId)
-	category, err := cmd.repo.GetCategory(categoryId)
-	if err != nil {
-		fmt.Println(err)
+	categoryId, readCatIdErr := cmd.ioUtil.ReadNumber()
+	if(readCatIdErr != nil) {
+		fmt.Println(readCatIdErr)
+		return
+	}
+	category, getCatErr := cmd.repo.GetCategory(categoryId)
+	if getCatErr != nil {
+		fmt.Println(getCatErr)
+		return
 	}
 	fmt.Println("Category details : ")
 	fmt.Println(category)
 }
-func (cmd CategoryCommand) EditCategory() {
+func (cmd *CategoryCommand) EditCategory() {
 	fmt.Println("Edit Category Command ...")
 	category := model.Category{}
 	fmt.Println("Please enter the id of the category : ")
-	fmt.Scanf("%d", &category.Id)
-
+	categoryId, readCatIdErr := cmd.ioUtil.ReadNumber()
+	if readCatIdErr != nil {
+		fmt.Println(readCatIdErr)
+		return
+	}
+	category.Id = categoryId
 	fmt.Println("Now enter the title of category : ")
-	fmt.Scanln(&category.Title)
-
+	categoryTitle, readCatTitleErr := cmd.ioUtil.ReadStr()
+	if readCatTitleErr != nil {
+		fmt.Println(readCatTitleErr)
+		return
+	}
+	category.Title = categoryTitle
 	fmt.Println("Now enter the color of category :")
-	fmt.Scanln(&category.Color)
-
+	categoryColor, readCatColorErr := cmd.ioUtil.ReadStr()
+	if readCatColorErr != nil {
+		fmt.Println(readCatColorErr)
+		return
+	}
+	category.Color = categoryColor
 	err := cmd.repo.EditCategory(category)	
 	if err != nil {
 		fmt.Println(err)
@@ -73,12 +102,14 @@ func (cmd CategoryCommand) EditCategory() {
 	}
 	fmt.Println("Category edited successfully")
 }
-func (cmd CategoryCommand) DeleteCategory() {
+func (cmd *CategoryCommand) DeleteCategory() {
 	fmt.Println("Delete Category Command ...")
 	fmt.Println("Please enter the id of the category : ")
-	var categoryId int
-	fmt.Scanf("%d", &categoryId)
-
+	categoryId, readCatIdErr := cmd.ioUtil.ReadNumber()
+	if readCatIdErr != nil {
+		fmt.Println(readCatIdErr)
+		return
+	}
 	err := cmd.repo.DeleteCategory(categoryId)
 	if err != nil {
 		fmt.Println(err)
